cxx-qt-gen: split parsed mappings into a separate file
