+++ crates/cxx-qt-gen/src/generator/cpp/method.rs
@@ -15,7 +15,7 @@ use crate::{
         },
     },
     parser::{
-        cxxqtdata::ParsedCxxMappings,
+        mappings::ParsedCxxMappings,
         method::{ParsedMethod, ParsedQInvokableSpecifiers},
     },
 };
