+++ crates/cxx-qt-gen/src/parser/cxxqtdata.rs
@@ -9,49 +9,18 @@ use crate::syntax::path::path_from_idents;
 use crate::syntax::safety::Safety;
 use crate::{
     parser::{
-        externcxxqt::ParsedExternCxxQt, inherit::ParsedInheritedMethod, method::ParsedMethod,
-        qobject::ParsedQObject, signals::ParsedSignal,
+        externcxxqt::ParsedExternCxxQt, inherit::ParsedInheritedMethod,
+        mappings::ParsedCxxMappings, method::ParsedMethod, qobject::ParsedQObject,
+        signals::ParsedSignal,
     },
     syntax::expr::expr_to_string,
 };
 use quote::format_ident;
 use std::collections::BTreeMap;
 use syn::{
-    spanned::Spanned, Attribute, Error, ForeignItem, Ident, Item, ItemForeignMod, ItemImpl, Path,
-    Result, Type, TypePath,
+    spanned::Spanned, Attribute, Error, ForeignItem, Ident, Item, ItemForeignMod, ItemImpl, Result,
+    Type, TypePath,
 };
-
-#[derive(Default)]
-pub struct ParsedCxxMappings {
-    /// Map of the cxx_name of any types defined in CXX extern blocks
-    ///
-    /// This is used in the C++ generation to map the Rust type name to the C++ name
-    pub cxx_names: BTreeMap<String, String>,
-    /// Map of the namespace of any types or methods defined in CXX extern blocks
-    ///
-    /// This is used in the C++ generation to map the Rust type name to the C++ name
-    pub namespaces: BTreeMap<String, String>,
-    /// Mappings for CXX types when used outside the bridge
-    pub qualified: BTreeMap<Ident, Path>,
-}
-
-impl ParsedCxxMappings {
-    /// For a given rust ident return the CXX name with its namespace
-    pub fn cxx(&self, ident: &str) -> String {
-        // Check if there is a cxx_name or namespace to handle
-        let cxx_name = self
-            .cxx_names
-            .get(ident)
-            .cloned()
-            .unwrap_or_else(|| ident.to_owned());
-
-        if let Some(namespace) = self.namespaces.get(ident) {
-            format!("::{namespace}::{cxx_name}")
-        } else {
-            cxx_name
-        }
-    }
-}
 
 pub struct ParsedCxxQtData {
     /// Mappings for CXX types when used in C++ or Rust
